forecaster/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forecaster"
version = "0.1.0"
description = "OptiPilot predictive autoscaling forecaster service"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.80.0",
    "grpcio-tools>=1.80.0",
    "protobuf>=5.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "aiosqlite>=0.20.0",
    "lightgbm>=4.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "joblib>=1.3",
]

[project.scripts]
forecaster = "forecaster.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.hatch.build.targets.wheel]
packages = ["src/forecaster"]

[tool.pytest.ini_options]
pythonpath = ["../gen/python"]
asyncio_mode = "auto"
testpaths = ["tests"]
